+++ src/base/b-dynamic-page/CHANGELOG.md
@@ -13,7 +13,7 @@ Changelog
 
 #### :bug: Bug Fix
 
-* Fixed a bug when a getter to current page component is `undefined`. Now the getter can return a promise.
+* Fixed a bug when the `component` getter is `undefined`. Now the getter can return a promise in that case.
 
 ## v3.0.0-rc.204 (2021-06-23)